libnetplan: expose the link-local setting
The low level API has two new functions that will return a boolean
representing the internal state of the property. In the netdef, this
property is stored in an internal struct that contains two gboolean
variables. To avoid having to expose this struct I decided to separate
it in two functions.

In the Python code, the new propserty 'link_local' will return a list
containing ipv4 or ipv6 or an empty list to mimic the link-local
property in the YAML.