ATTN: parse/bonds: handle same primary in multiple bonds
Error out if an interface is declared as primary of multiple bonds.
Besides being wrong it's causing a memory leak.

ATTN: as this makes the parser stricter, it probably should not be
included in backports/SRUs. Consider converting the error to a warning
and freeing the memory allocation to avoid the leak.