sriov: accept setting the eswitch mode without VFs
The embedded switch mode can be set even if the interface doesn't have
any virtual functions. This is a requisite to use the PF with only
scalable functions for example.

With this change, the presence of the 'embedded-switch-mode' will be
enough to identify the interface as a PF and emit the systemd services
required to apply the configuration.

The changes required by the CLI will be done is the next commit.