cli/sriov: refactoring
Refactor get_vf_count_and_functions() in 3 different ones. It was doing
too many things and was hard to read. Three of its 5 parameters were being
used as output and one of them was being changed by
_get_target_interface() in the same call chain.

The new _get_physical_functions() will also return interfaces that only
have the property embedded_switch_mode. This is a requirement to enable
netplan to change the eswitch mode even if it doesn't have VFs.
See LP: #2020409

Adapt the unit tests to the new functions and implement new tests.