ATTN: parse/bonds: handle same primary in multiple bonds (#451)
* ATTN: parse/bonds: handle same primary in multiple bonds

Error out if an interface is declared as primary of multiple bonds.
Besides being wrong it's causing a memory leak.

ATTN: as this makes the parser stricter, it probably should not be
included in backports/SRUs. Consider converting the error to a warning
and freeing the memory allocation to avoid the leak.

* parse/bond: identify bond assigned to primary

Look for the bond that already has the primary member assigned and
identify it in the error message.