generate: enable systemd-networkd-wait-online for non-optional interfaces only
Skip s-n-wait-online if we don't have any non-optional interfaces, using a
"ConditionPathIsSymbolicLink=" checking Netplan's s-n-wait-online.service
enablement symlink.

This is in favor to RequiredForOnline=yes as the behavior of upstream (pure)
systemd-networkd-wait-online.service is not mean to be used in this way.

If "RequiredForOnline=no" sd-networkd-wait-online will fully ignore the
corresponding interface and it will block/delay network-online.target if
no interfaces are "RequiredForOnline=yes" at all.

FR-7246