parse: fix redefinition of gateway(4|6)
Options gateway4 and gateway6 were not being properly redefined when a
merge happens. Redefinition was causing a crash due to an assert in the
set_str_if_null macro.

While this assertion helped identifying the merge issue, I believe it's
not the best place to keep it so I'm removing it from the macro and
making the macro do only what it name suggests it does.

This macro is used in other places but this problem seems to not happen
there as far as I can see.

Add a couple of unit test to check the merge is working.