generate: add support for ignoring errors
Add a new flag --ignore-errors. It it's used mostly for tests.

IGNORE_ERRORS will be set either if --ignore-errors or the environment
variable NETPLAN_PARSER_IGNORE_ERRORS are set. These are mostly used for
tests from the unit and integration tests.

The flag will be set by default if "generate" is called as a systemd
generator. The idea is to ignore errors when the machine boots up so the
system still has a chance to have a working network configuration if an
issue is introduced in any of the YAML files.

For the time being, this flag will not be enabled by default when the
CLI is used. This will make the issue evident so the user will need to
fix it.