parse: introduce parser flags
Parser flags are intended to be used to change the parser's behavior.
This patch introduces a single flag used to instruct the parser that
parsing errors should be ignored and it should continue to parse
netdefs.

The main goal with this flag is to make Netplan more resilient.
Currently, any mistake made in the YAML files will prevent Netplan to
generate configuration. So, rebooting the system while one of the YAML
files are broken will result in the system having no network connectivity.