parse: implement the IGNORE_ERRORS flag
When the ignore_errors flag is enabled, the parser will ignore YAML
files that failed to be loaded due to syntax errors. YAMLs that are
loaded successfully but contain invalid Netplan configuration will not
be completely ignored.

Network definitions that contain errors will not be completely ignored.
The parser will ignore the error in the bad YAML mapping and continue
with the next one. By doing it, all the good definitions will produce
backend configuration and the bad ones will produce *some* backend
configuration.

Add a simple counter to the parser struct that will store the number of
errors that were ignored. This can be used by callers to check if there
were errors when the IGNORE_ERRORS flag is used.