state_diff: add gateway(4|6) to the routes list
The (deprecated but still available) properties gateway4 and gateway6
were not being taken into account when comparing routes. Because of
that, the diff report will show the default route as missing in netplan.

There is a complication when considering gateway4 and gateway6 though:
both networkd and NM will set a metric for them. As these properties
cannot set metrics, these routes would still be shown in the diff
report.

To avoid this noise, when gateway4 and/or gateway6 are used AND the
system contains one, and only one, default route per protocol family, we
adopt the same metric and route table from the existing routes. By doing
this, the comparison will see that both routes, from the system and from
netplan, are the same and not show them as a diff.