state: add the 'ra' flag to IPv6 addresses
There are cases where IPv6 addresses received via RA are not set as
'dynamic'. In these cases, use the information about routes to find
these IPs and add the tag 'ra' to them.

It's done by iterating through IPv6 addresses and checking if there is
any addresses in the same network as routes received via RA.