api: add a getter for accept-ra
Add the Python binding and unit test.

NOTE

accept_ra is not perfectly mapped from the netdef structure to the
YAML setting and back to the renderer. It's a boolean value in the YAML
but it maps to three different values inside libnetplan: kernel, enabled
and disabled.

Not setting it in the YAML means it will not be set in
networkd. Although, networkd will disable it unconditionally in
the kernel and will itself handle RAs, meaning that when unset, it's
enabled by default (with a few exceptions, see systemd.network(5)).

Because of that, the low level C API will return the actual internal
value of the property and not only TRUE or FALSE. The Python code will
return True, False or None (when it's unset). This will help the
consumers to differentiate the case where it's disabled or unset.