state_diff: improve analysis of RA/LL addresses
When IPv6 link-local and RA (which depends on link-local) are enabled,
the system will receive IP addresses, nameservers and routes dynamically
via RA. These addresses are not obviously identified as dynamic in the
system. networkd probably does have this information but it's not fully
exposed apparently.

Here I try to improve the comparison of these addresses by checking if
they are 1) a link-local address and 2) part of the same network as the
route received via RA.

Without this, status --diff will show the RA IP addresses and link-local
DNS nameserver as a diff.