state: make use of networkd ConfigSource
When networkd is used, the ConfigSource property can identify the source
of the configuration so let's use it to be more precise and reduce the
guessing.

We consume the same information when Network Manager is the renderer but
in this case networkd will not know the sources so this property can't
be relied on.

For now, we collect the sources for IP addresses and DNS. Routes can be
included in the future. We also should create abstractions for each
configuration so we don't need to keep information such as the data
source in separate tables.

Network Manager seems to expose the source through its DBus API. This
also could be considered in the future.