wait-online: disabled wait-online for stable 1.0
The new wait-online feature introduces an important change in behavior
and the generation of the config file and systemd unit is disabled.

Keeping the code will help to reduce the number of distro patches
required when merging netplan.io 1.0.1. The only patch required for
wait-online will be the one reverting this commit.