parse: do not escape all non-ascii bytes
g_strescape will transform utf-8 strings in a sequence of escaped octal
numbers in form of a string. That happens because g_strescape will
escape the range from 0x7f to 0xff.

For example, "áéí" will become "\\303\\241\\303\\251\\303\\255". Pure
ASCII strings are not affected by it.

Add all the range 0x7f-0xff to the exception list.