Change default umask when creating dirctories
Security best practices recommend changing the default umask from 022
to 027 to harden systems. However, when users set umask to 027,
netplan creates directories with permissions that are not accessible
to backends like systemd-network, resulting in issues during network
configuration.

This patch ensures that netplan generates directories with a 0755
mask, regardless of the user's umask.

fixes: lp2076319