apply: don't skip members of bonds on renaming
Skipping members of bonds when searching for interfaces that need to
be renamed was done as a workaround for LP: #1802322.

The actual problem was that members of bonds will have their MAC
addresses changed to a common address when they are added as members.
The renaming process looked for interfaces based on their transient MAC
addresses not the permanent one. Because of that, it would confuse one
interface by another and try to rename it.

The process was changed to look for the permanent MAC address so this
problem shouldn't happen. It also searches only for physical interfaces,
which should have a permanent MAC address.

Skipping the renaming of members of bonds will cause the backend to
"forget" about it. Once the configuration is generated with a different
name, the backend will look for the new name. As the interface will
continue to have the previous name, it will not be managed anymore.

Lets fix this by allowing the renaming these interfaces. If the system
administrator changed the "set-name" setting, they probably want to get
the interface renamed. This might cause a quick loss of connectivity on
the interface being renamed.

Drop is_composite_member(). This method is not needed anymore.