state/Interface: improve netdef ID handling with NetworkManager
The ID is extracted from the file name. For Wi-Fi connections, the SSID
will also be present in the file name and we need to remove it in order
to extract the netdef ID. When non-ascii characters are
present in the SSID, NM will represent it using a byte sequence
separated with ;. The semicolons are escaped in the file name when we
emit the .nmconnection file.

In order to find and remove the SSID from the file name, so we can get
the netdef ID, we need to convert it to the same representation.

To detect this situation we check if the SSID is present in the file
name. If it's not, it likely because the SSID contains non-ascii
characters.