parse: new structure for "passthrough"
The networkmanager.passthrough section stores raw keyfile information
using a format that can be tricky to interpret. It will concatenate the
group and key names separated by dots. Unfortunately, Network Manager
accepts multiple dots in the middle of both the group and key names.
Because of that it's hard, if even possible, to determine where the
strings must be split.

Ex:
networkmanager:
  passthrough:
    group_name.key_name: value1
    group_name.key_name1: value2

The new format stores the same information in YAML mappings. By using
this format, parsing the group and key names is not needed anymore.

Ex:
networkmanager:
  passthrough:
    group_name:
      key_name1: value1
      key_name2: value2

The datalist data structure was replaced by hash tables. The entire
passthrough section is represented by a hash table indexed by the group
name. Each entry is also a hash table representing all the key/value
entries and it's indexed by the key name.