parse: fix renderer validation for nm-devices
When we process a nm-device we check to see if the renderer is
NetworkManager. If it's not set to NetworkManager we log a warning
message and change the backend.

This verification can be only done after the netdef is processed. Move
the verification so it happens after the call to process_mapping().

Fixes LP: #2091755