Allows non standard OVS setups
Changes the checks to be based around the availability of the command
line OVS tools instead of the availability of the service installed
under apt. This allows netplan to consume OVS when it is provided as part
of another package such as in snap environments.

This involves minor changes to lots of the OVS code and also modifies
the unit tests as to handle OVS from both standard and nonstandard setups.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>