configure: drop JIT logic, which is not needed anymore with the sd-generator
This was originally implemented to generate & start systemd units just-in-time
during the boot transaction (https://github.com/canonical/netplan/pull/162).

With the implementation of a proper systemd-generator, Netplan generates the
corresponding units in /run/system/generator* and automatically re-loads and
re-calculates dependencies during "daemon-reload". Therefore, we do not need
to inject them manually.

This is covered by the "cloud-init" autopkgtest.